:white_check_mark: apply search-bar feature
