+++ front-end/books/src/components/MainPage/BookListPage/BookList.js
@@ -1,42 +1,21 @@
 // import qs from "qs";
-import {useState, useEffect} from "react"
-import axios from "axios"
 
-const BookList = ({location, match, history}) => {
-  const [books, setBooks] = useState(false);
-  const {inputs, tag} = location.state;
-
-  useEffect(() => {
-    const getBooklist = async () => {
-      await axios.get("http://localhost:8080/api/items/", {
-        params: {
-          name:inputs,
-        }
-      })
-        .then(res => {
-          console.log(res.data._embedded.itemDtoList)
-          setBooks(res.data._embedded.itemDtoList)
-        })
-        .catch(err => {
-          console.log(err.response)
-          setBooks(false)
-        })
-    }
-    getBooklist()
-  }, [inputs])
-
+const BookList = ({location}) => {
+  const {input, tag, books} = location.state
+  console.log(location)
   return <>
     <div>태그명 : {tag} </div>
-    <div>검색어 : {inputs}</div>
-    {!books ? 
-    <div>검색 결과가 없습니다.</div> 
-    :
-    books.map((book, idx) => {
-      return <div key={idx}>
-        <h2>{book.name}</h2>
-        <p>{book.description}</p>
-      </div>
-    })}
+    <div>검색어 : {input}</div>
+    {books ?
+      books.map((book, idx) => {
+        return <div key={idx}>
+          <h2>{book.name}</h2>
+          <p>{book.description}</p>
+        </div>
+      })
+      :
+      <div>검색 결과가 없습니다.</div>
+    }
   </>
 }
 export default BookList;
